CDP: fix CHECK failure in Page.resetNavigationHistory
Fixed: 1513335
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5145010
Auto-Submit: Yang Guo <yangguo@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1240359}

+++ content/browser/devtools/protocol/page_handler.cc
@@ -772,8 +772,12 @@ Response PageHandler::ResetNavigationHistory() {
     return response;
 
   NavigationController& controller = host_->frame_tree()->controller();
-  controller.DeleteNavigationEntries(base::BindRepeating(&ReturnTrue));
-  return Response::Success();
+  if (controller.CanPruneAllButLastCommitted()) {
+    controller.DeleteNavigationEntries(base::BindRepeating(&ReturnTrue));
+    return Response::Success();
+  } else {
+    return Response::ServerError("History cannot be pruned");
+  }
 }
 
 void PageHandler::CaptureSnapshot(
